refactor(gui): improve file structure
