+++ api-editor/gui/src/common/util/stringOperations.ts
@@ -1,3 +1,7 @@
 export const truncate = function (text: string, maxLength: number): string {
     return text.length > maxLength ? text.substring(0, maxLength - 1) + '\u2026' : text;
 };
+
+export const pluralize = function (count: number, noun: string): string {
+    return `${count} ${noun}${count === 1 ? '' : 's'}`;
+};
