+++ api-editor/gui/src/features/achievements/AchievementCard.tsx
@@ -1,0 +1,28 @@
+import { Center, Img, Text as ChakraText, VStack } from '@chakra-ui/react';
+import React from 'react';
+import { Achievement } from './achievements';
+
+interface AchievementCardProps {
+    achievement: Achievement;
+    currentCount: number;
+    description: string;
+}
+
+export const AchievementCard: React.FC<AchievementCardProps> = function ({ achievement, currentCount, description }) {
+    const currentAchievementLevel = achievement.level(currentCount);
+    if (!currentAchievementLevel) {
+        return null;
+    }
+
+    return (
+        <VStack boxShadow="xl" rounded="md" border={1} layerStyle="subtleBorder" padding={4}>
+            <Img src={currentAchievementLevel.image} alt={currentAchievementLevel.label} />
+            <Center fontWeight="bold">
+                <ChakraText>{currentAchievementLevel.label}</ChakraText>
+            </Center>
+            <Center>
+                <ChakraText textAlign="center">{description}</ChakraText>
+            </Center>
+        </VStack>
+    );
+};
